test_vercel_deployment.py
        from agent.registry import AgentRegistry
        from fastapi.middleware.cors import CORSMiddleware
        from fastapi.responses import JSONResponse
        from ml.model_registry import ModelRegistry
        from ml.recommendation_engine import RecommendationEngine
        from sqlalchemy import create_engine
        import secrets
import sys

        from fastapi import FastAPI, HTTPException, Request
        from pydantic import BaseModel, EmailStr
        from pydantic import BaseModel, Field
        import sqlalchemy

        from agent.modules.backend.fixer_v2 import FixerAgentV2
        from agent.modules.base_agent import BaseAgent
        from agent.modules.frontend.web_development_agent import WebDevelopmentAgent
        from agent.orchestrator import AgentOrchestrator
        from cryptography.fernet import Fernet
        from fashion.intelligence_engine import FashionIntelligenceEngine
        from typing import Optional
        from vercel_startup import validate_email_dependencies, get_app_config
        import anthropic
        import email_validator
        import hashlib
        import main
        import openai
        import uvicorn
from typing import Dict, List, Tuple
import logging
import traceback

logger = logging.getLogger(__name__)
#!/usr/bin/env python3
"""
DevSkyy Vercel Deployment Test Suite v1.0.0

Comprehensive test suite to validate Vercel deployment readiness.
Tests all critical dependencies, imports, and functionality.

Author: DevSkyy Team
Version: 1.0.0
Python: >=3.11
"""

def test_email_validation() -> Tuple[bool, str]:
    """Test email validation functionality."""
    try:
        # Test 1: Import email-validator directly

        # Test 2: Test Pydantic email validation

        class TestModel(BaseModel):
            email: EmailStr

        # Test 3: Validate an email
        test_instance = TestModel(email="test@example.com")

        return True, "✅ Email validation working correctly"

    except ImportError as e:
        return False, f"❌ Email validation import failed: {e}"
    except Exception as e:
        return False, f"❌ Email validation test failed: {e}"

def test_pydantic_functionality() -> Tuple[bool, str]:
    """Test Pydantic core functionality."""
    try:

        class TestModel(BaseModel):
            name: str
            age: int = Field(ge=0, le=150)
            email: Optional[str] = None

        # Test model creation
        instance = TestModel(name="Test User", age=25, email="test@example.com")

        # Test validation
        try:
            TestModel(name="Test", age=-1)  # Should fail
            return False, "❌ Pydantic validation not working"
        except Exception as e:
    logger.warning(f"Handled exception: {e}")  # Expected to fail

        return True, "✅ Pydantic functionality working correctly"

    except Exception as e:
        return False, f"❌ Pydantic test failed: {e}"

def test_fastapi_imports() -> Tuple[bool, str]:
    """Test FastAPI and related imports."""
    try:

        # Test FastAPI app creation
        app = FastAPI(title="Test App")

        return True, "✅ FastAPI imports working correctly"

    except Exception as e:
        return False, f"❌ FastAPI import test failed: {e}"

def test_core_modules() -> Tuple[bool, str]:
    """Test core application modules."""
    try:
        # Test agent modules

        # Test ML modules

        return True, "✅ Core modules loading correctly"

    except Exception as e:
        return False, f"❌ Core modules test failed: {e}"

def test_optional_modules() -> Tuple[bool, str]:
    """Test optional modules with graceful degradation."""
    results = []

    # Test V2 Backend Agents
    try:

        results.append("✅ V2 Backend agents available")
    except ImportError:
        results.append("ℹ️ V2 Backend agents not available (optional)")

    # Test V2 Frontend Agents
    try:

        results.append("✅ V2 Frontend agents available")
    except ImportError:
        results.append("ℹ️ V2 Frontend agents not available (optional)")

    # Test Fashion Intelligence
    try:

        results.append("✅ Fashion intelligence engine available")
    except ImportError:
        results.append("ℹ️ Fashion intelligence engine not available (optional)")

    # Test AI clients
    ai_clients = []
    try:

        ai_clients.append("Anthropic")
    except ImportError as e:
    logger.warning(f"Handled exception: {e}")

    try:

        ai_clients.append("OpenAI")
    except ImportError as e:
    logger.warning(f"Handled exception: {e}")

    if ai_clients:
        results.append(f"✅ AI clients available: {', '.join(ai_clients)}")
    else:
        results.append("ℹ️ No AI clients available (optional)")

    return True, "\n".join(results)

def test_database_functionality() -> Tuple[bool, str]:
    """Test database functionality."""
    try:

        # Test SQLite connection (default for Vercel)
        engine = create_engine("sqlite:///:memory:")
        connection = engine.connect()
        connection.close()

        return True, "✅ Database functionality working"

    except Exception as e:
        return False, f"❌ Database test failed: {e}"

def test_security_modules() -> Tuple[bool, str]:
    """Test security and encryption modules."""
    try:

        # Test encryption
        key = Fernet.generate_key()
        cipher = Fernet(key)

        # Test token generation
        token = secrets.token_urlsafe(32)

        # Test hashing
        hash_value = hashlib.sha256(b"test").hexdigest()

        return True, "✅ Security modules working correctly"

    except Exception as e:
        return False, f"❌ Security test failed: {e}"

def test_vercel_startup() -> Tuple[bool, str]:
    """Test Vercel startup module."""
    try:

        # Test email validation check
        email_available = validate_email_dependencies()

        # Test app config
        config = get_app_config()

        return True, f"✅ Vercel startup working (email validation: {email_available})"

    except Exception as e:
        return False, f"❌ Vercel startup test failed: {e}"

def test_main_application() -> Tuple[bool, str]:
    """Test main application import."""
    try:
        # This should not raise any import errors

        return True, "✅ Main application imports successfully"

    except Exception as e:
        return False, f"❌ Main application test failed: {e}"

def run_all_tests() -> Dict[str, Tuple[bool, str]]:
    """Run all deployment tests."""
    tests = {
        "Email Validation": test_email_validation,
        "Pydantic Functionality": test_pydantic_functionality,
        "FastAPI Imports": test_fastapi_imports,
        "Core Modules": test_core_modules,
        "Optional Modules": test_optional_modules,
        "Database Functionality": test_database_functionality,
        "Security Modules": test_security_modules,
        "Vercel Startup": test_vercel_startup,
        "Main Application": test_main_application,
    }

    results = {}

    print("🧪 Running DevSkyy Vercel Deployment Tests...")
    print("=" * 60)

    for test_name, test_func in tests.items():
        print(f"\n🔍 Testing: {test_name}")
        try:
            success, message = test_func()
            results[test_name] = (success, message)
            print(f"   {message}")
        except Exception as e:
            error_msg = f"❌ Test execution failed: {e}"
            results[test_name] = (False, error_msg)
            print(f"   {error_msg}")
            print(f"   Traceback: {traceback.format_exc()}")

    return results

def print_summary(results: Dict[str, Tuple[bool, str]]):
    """Print test summary."""
    print("\n" + "=" * 60)
    print("📊 TEST SUMMARY")
    print("=" * 60)

    passed = sum(1 for success, _ in results.values() if success)
    total = len(results)

    print(f"Tests Passed: {passed}/{total}")
    print(f"Success Rate: {(passed/total)*100:.1f}%")

    print("\n📋 DETAILED RESULTS:")
    for test_name, (success, message) in results.items():
        status = "✅ PASS" if success else "❌ FAIL"
        print(f"  {status} - {test_name}")
        if not success:
            print(f"    └─ {message}")

    print("\n🎯 DEPLOYMENT READINESS:")
    critical_tests = [
        "Pydantic Functionality",
        "FastAPI Imports",
        "Core Modules",
        "Main Application",
    ]

    critical_passed = sum()
        1 for test in critical_tests if results.get(test, (False, ""))[0]
    )

    if critical_passed == len(critical_tests):
        print("✅ READY FOR DEPLOYMENT")
        print("   All critical tests passed. Application should deploy successfully.")
    else:
        print("❌ NOT READY FOR DEPLOYMENT")
        print("   Critical tests failed. Fix issues before deploying.")

        failed_critical = [
            test for test in critical_tests if not results.get(test, (False, ""))[0]
        ]
        print(f"   Failed critical tests: {', '.join(failed_critical)}")

    # Optional features summary
    optional_tests = [
        "Email Validation",
        "Optional Modules",
        "Database Functionality",
        "Security Modules",
    ]

    optional_passed = sum()
        1 for test in optional_tests if results.get(test, (False, ""))[0]
    )
    print(f"\n📈 Optional Features: {optional_passed}/{len(optional_tests)} available")

if __name__ == "__main__":
    print("🚀 DevSkyy Vercel Deployment Test Suite")
    print("Testing deployment readiness and dependency validation...")

    results = run_all_tests()
    print_summary(results)

    # Exit with appropriate code
    critical_tests = [
        "Pydantic Functionality",
        "FastAPI Imports",
        "Core Modules",
        "Main Application",
    ]
    critical_passed = sum()
        1 for test in critical_tests if results.get(test, (False, ""))[0]
    )

    if critical_passed == len(critical_tests):
        print("\n🎉 All critical tests passed! Ready for Vercel deployment.")
        sys.exit(0)
    else:
        print("\n💥 Critical tests failed! Fix issues before deployment.")
        sys.exit(1)
